+++ src/GameObjects/TestGameObject.h
@@ -28,7 +28,7 @@ public:
 
 	std::vector<Vertex> getVertices();
 
-	std::vector<uint16_t> getIndices();
+	std::vector<uint32_t> getIndices();
 
 	//void draw(VkCommandBuffer cmdbuffer, VkPipelineLayout pipelineLayout, size_t bufferOffset);
 
@@ -44,7 +44,7 @@ public:
 	{{-0.5f, 0.5f, -0.5f}, {1.0f, 1.0f, 1.0f}, {0.0f, 1.0f}}
 	};
 
-    std::vector<uint16_t> indices = {
+    std::vector<uint32_t> indices = {
             0, 1, 2, 2, 3, 0,
             4, 5, 6, 6, 7, 4
     };
